cpuidle: Avoid crashing in cpuidle_driver_state_disabled()
The cpumask is NULL if cpuidle_driver_state_disabled() is called before
cpuidle driver is initialized. This shouldn't happen for anyone for now
because cpuidle drivers are registered quite early, but it will be a
problem for cpuidle drivers that probe later during kernel's boot-up.
The NVIDIA Tegra cpuidle driver is one of those drivers and it's going
to be moved to a later init stage.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>